requirements.txt
KNN_CUDA==0.2
matplotlib==3.5.3
numpy==1.21.5
open3d==0.17.0
pandas==1.3.5
Pillow==9.5.0
Pillow==10.3.0
pointnet2_ops==3.0.0
scikit_learn==1.0.2
scipy==1.7.3
tifffile==2021.11.2
timm==0.4.5
torch==1.10.0
torchvision==0.11.1
tqdm==4.66.1
trimesh==4.0.5
